requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-import-export==2.6.1
django-rest-framework==0.1.0
django-simple-history==3.0.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
numpy==1.21.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
psycopg2==2.9.2
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.2.4
tablib==3.1.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
